2022.12.10 42

+++ libft/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: danlopez <danlopez@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/06 06:35:19 by danlopez          #+#    #+#             */
-/*   Updated: 2022/12/06 06:44:54 by danlopez         ###   ########.fr       */
+/*   Updated: 2022/12/10 11:41:09 by danlopez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
 	int	len;
